+++ src/main/java/ch/zhaw/schiffeversenken/data/Ship.java
@@ -3,8 +3,7 @@ package ch.zhaw.schiffeversenken.data;
 import java.util.ArrayList;
 import java.util.List;
 
-import ch.zhaw.schiffeversenken.helpers.Coordinate;
-import ch.zhaw.schiffeversenken.helpers.Directions;
+import ch.zhaw.schiffeversenken.model.Directions;
 
 /**
  * Simple Ship representation. Has only a position (list of coordinates). The
