+++ src/main/java/ch/zhaw/schiffeversenken/gamestart/GameStarter.java
@@ -3,7 +3,7 @@ package ch.zhaw.schiffeversenken.gamestart;
 import javax.swing.UIManager;
 import javax.swing.UnsupportedLookAndFeelException;
 
-import ch.zhaw.schiffeversenken.guicomponents.StartScreen01;
+import ch.zhaw.schiffeversenken.view.StartScreen01;
 
 public class GameStarter {
 	public static void main(String[] args) throws ClassNotFoundException, InstantiationException, IllegalAccessException, UnsupportedLookAndFeelException {
