+++ src/test/java/ch/zhaw/schiffeversenken/data/PlayFieldTest.java
@@ -9,8 +9,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-import ch.zhaw.schiffeversenken.helpers.Coordinate;
-import ch.zhaw.schiffeversenken.helpers.Directions;
+import ch.zhaw.schiffeversenken.model.Directions;
 
 public class PlayFieldTest {
 	PlayField playField;
